requirements.txt
Flask>=2.3.0
tensorflow>=2.20.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
Werkzeug>=2.3.0
scikit-learn>=1.3.0
gunicorn>=21.0.0
